Read Abilities from YAML
Added functionality to read abilities from YAML files. I'm not sure how much I like it because it shuts down some of the more personal aspects of hard-coding each ability in, such as the ability to retain ability usage between CharacterPage instances.

I'll continue to play with it and see if I can increase that functionality to my satisfaction.